Refactor: Split orchestration up into manageable chunks
